backend/requirements/base.txt
# ─── Django core ───────────────────────────────────────────────────────
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
python-decouple==3.8

# ─── Wagtail CMS ───────────────────────────────────────────────────────
wagtail==6.3.1
wagtail-headless-preview==0.8.0
wagtail-modeladmin==2.2.0

# ─── Auth ──────────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ─── API docs ──────────────────────────────────────────────────────────
drf-spectacular==0.27.2

# ─── Database ──────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ─── Cache / Queue ─────────────────────────────────────────────────────
redis==5.0.8
django-redis==5.4.0
celery==5.4.0

# ─── Images ────────────────────────────────────────────────────────────
Pillow==10.4.0
willow==1.8.0

# ─── Utilities ─────────────────────────────────────────────────────────
django-extensions==3.2.3
django-filter==24.3